minor cabal change
