+++ hgit.cabal
@@ -51,9 +51,14 @@ library
       CommandParser
       CommandHandler
       Commit
-      FileIO
-      Hash
       Index
+      Utils
+      TestUtils
+      AddTests
+      CommitTests
+      HashTests
+      InitTests
+      ParserTests
       -- if you add modules to your project, list them here
       -- so that they will be available to the executable and test suite
   hs-source-dirs:
@@ -80,13 +85,6 @@ test-suite hgit-test
       Spec.hs
   hs-source-dirs:
       test
- other-modules:
-      Parser,
-      Commit,
-      Add,
-      Hash,
-      Init,
-      Utils
   ghc-options: -threaded -rtsopts -with-rtsopts=-N
   build-depends:
       hgit
