+++ src/Index.hs
@@ -3,30 +3,24 @@ module Index
     writeIndexFile,
     updateIndex,
     getAllFiles,
-    getTrackedFiles
-  ) where
+    getTrackedFiles,
+  )
+where
 
-import CommandParser (CommandError(..))
-import FileIO
-    ( doesDirectoryExist,
-      createDirectoryIfMissing',
-      getObjectsPath,
-      getIndexFilePath,
-      createObject )
-import Hash
-import System.Directory (doesFileExist, listDirectory, getCurrentDirectory, doesDirectoryExist)
-import System.FilePath ((</>), normalise, makeRelative)
 import Control.Monad (filterM, forM, unless)
-import Data.Map.Strict (Map)
-import qualified Data.Map.Strict as Map
 import Data.ByteString (ByteString)
-import qualified Data.ByteString as BS
-import qualified Data.ByteString.Char8 as BS8
+import Data.ByteString qualified as BS
+import Data.ByteString.Char8 qualified as BS8
+import Data.Map.Strict (Map)
+import Data.Map.Strict qualified as Map
 import Data.Maybe (catMaybes)
 import Data.Text.Encoding.Error (UnicodeException)
-
+import System.Directory (doesDirectoryExist, doesFileExist, getCurrentDirectory, listDirectory)
+import System.FilePath (makeRelative, normalise, (</>))
+import CommandParser
+import Utils
 -- Type alias for the index mapping
-type IndexMap = Map FilePath String  -- Map from file path to OID
+type IndexMap = Map FilePath String -- Map from file path to OID
 
 -- | Reads the index file into a Map
 readIndexFile :: IO IndexMap
@@ -44,10 +38,10 @@ readIndexFile = do
 -- | Parses a line from the index file into a (FilePath, OID) tuple
 parseIndexLine :: ByteString -> (FilePath, String)
 parseIndexLine line =
-  let (oidBS, pathBS) = BS.break (== 32) line  -- 32 is ASCII code for space
+  let (oidBS, pathBS) = BS.break (== 32) line -- 32 is ASCII code for space
       oid = either (error "Invalid OID encoding") id $ byteStringToString oidBS
       path = either (error "Invalid path encoding") id $ byteStringToString (BS.drop 1 pathBS)
-  in (path, oid)
+   in (path, oid)
 
 -- | Writes the index Map back to the index file
 writeIndexFile :: IndexMap -> IO ()
@@ -77,11 +71,11 @@ updateTrackedFiles indexMap = do
 -- | Updates the index with the specified files
 updateFiles :: IndexMap -> [FilePath] -> IO (Either CommandError IndexMap)
 updateFiles indexMap files = do
-  results <- mapM (processFile indexMap) files  -- Now returns list of updates
+  results <- mapM (processFile indexMap) files -- Now returns list of updates
   case sequence results of
     Left err -> return $ Left err
     Right updatesList -> do
-      let updates = concat updatesList  -- Flatten list of lists
+      let updates = concat updatesList -- Flatten list of lists
       let newIndexMap = foldl (\accMap (fp, oid) -> Map.insert fp oid accMap) indexMap updates
       return $ Right newIndexMap
 
@@ -98,17 +92,18 @@ processFile indexMap filepath = do
         Left err -> return $ Left err
         Right maybeUpdate ->
           case maybeUpdate of
-            Nothing -> return $ Right []  -- No changes
-            Just update -> return $ Right [update]  -- Single file update
-    else if isDir
-      then do
-        -- Recursively process all files in the directory
-        files <- getFilesRecursive filepath
-        results <- mapM (processFileContents indexMap) files
-        case sequence results of
-          Left err -> return $ Left err
-          Right updates -> return $ Right $ catMaybes updates  -- Collect updates
-      else return $ Left $ CommandError $ "fatal: pathspec '" ++ filepath ++ "' did not match any files"
+            Nothing -> return $ Right [] -- No changes
+            Just update -> return $ Right [update] -- Single file update
+    else
+      if isDir
+        then do
+          -- Recursively process all files in the directory
+          files <- getFilesRecursive filepath
+          results <- mapM (processFileContents indexMap) files
+          case sequence results of
+            Left err -> return $ Left err
+            Right updates -> return $ Right $ catMaybes updates -- Collect updates
+        else return $ Left $ CommandError $ "fatal: pathspec '" ++ filepath ++ "' did not match any files"
 
 processFileContents :: IndexMap -> FilePath -> IO (Either CommandError (Maybe (FilePath, String)))
 processFileContents indexMap filepath = do
@@ -120,7 +115,7 @@ processFileContents indexMap filepath = do
   let normalizedPath = normalise relativePath
   let existingOid = Map.lookup normalizedPath indexMap
   if existingOid == Just oid
-    then return $ Right Nothing  -- No changes needed
+    then return $ Right Nothing -- No changes needed
     else do
       createObject content
       return $ Right $ Just (normalizedPath, oid)
@@ -133,7 +128,6 @@ getAllFiles = do
 
 -- | Helper function to get files recursively, excluding .hgit
 -- In Index.hs
-
 getFilesRecursive :: FilePath -> IO [FilePath]
 getFilesRecursive dir = do
   contents <- listDirectory dir
