--- test/Add.hs
@@ -1,1 +1,0 @@
-module Add where
