--- test/Commit.hs
@@ -1,1 +1,0 @@
-module Commit where
