--- test/Hash.hs
@@ -1,1 +1,0 @@
-module Hash where
