--- test/Init.hs
@@ -1,1 +1,0 @@
-module Init where
