--- test/Utils.hs
@@ -1,1 +1,0 @@
-module Utils where
